+++ src/ViewHoldingsRecord.js
@@ -65,6 +65,7 @@ import {
 } from './components';
 import HoldingAquisitions from './Holding/ViewHolding/HoldingAquisitions';
 import HoldingReceivingHistory from './Holding/ViewHolding/HoldingReceivingHistory';
+import HoldingBoundWith from './Holding/ViewHolding/HoldingBoundWith';
 
 import css from './View.css';
 
@@ -108,6 +109,16 @@ class ViewHoldingsRecord extends React.Component {
       DELETE: {
         path: 'source-storage/records/%{marcRecordId}',
       },
+    },
+    boundWithItems: {
+      type: 'okapi',
+      records: 'items',
+      path: 'inventory/items-by-holdings-id',
+      params: {
+        query: '(holdingsRecordId==:{holdingsrecordid})',
+        limit: '5000',
+      },
+      resourceShouldRefresh: true,
     },
   });
 
@@ -456,6 +467,7 @@ class ViewHoldingsRecord extends React.Component {
       resources: {
         items,
         tagSettings,
+        boundWithItems,
       },
       referenceTables,
       goTo,
@@ -468,8 +480,10 @@ class ViewHoldingsRecord extends React.Component {
     const instanceSource = referenceTables?.holdingsSources?.find(source => source.name === instance?.source);
     const holdingsRecord = this.getMostRecentHolding();
     const holdingsSource = referenceTables?.holdingsSources?.find(source => source.id === holdingsRecord.sourceId);
-    const holdingsPermanentLocation = get(referenceTables?.locationsById[holdingsRecord?.permanentLocationId], ['name'], '-');
-    const holdingsTemporaryLocation = get(referenceTables?.locationsById[holdingsRecord?.temporaryLocationId], ['name'], '-');
+    const holdingsPermanentLocation = referenceTables?.locationsById[holdingsRecord?.permanentLocationId];
+    const holdingsPermanentLocationName = get(holdingsPermanentLocation, ['name'], '-');
+    const holdingsTemporaryLocation = referenceTables?.locationsById[holdingsRecord?.temporaryLocationId];
+    const holdingsEffectiveLocation = referenceTables?.locationsById[holdingsRecord?.effectiveLocationId];
     const itemCount = get(items, 'records.length', 0);
     const holdingsSourceName = holdingsSource?.name || instanceSource?.name;
     const tagsEnabled = !tagSettings?.records?.length || tagSettings?.records?.[0]?.value === 'true';
@@ -479,7 +493,7 @@ class ViewHoldingsRecord extends React.Component {
         id="ui-inventory.confirmHoldingsRecordDeleteModal.message"
         values={{
           hrid: holdingsRecord.hrid,
-          location: holdingsPermanentLocation
+          location: holdingsPermanentLocationName
         }}
       />
     );
@@ -493,7 +507,7 @@ class ViewHoldingsRecord extends React.Component {
         id={noHoldingsRecordDeleteModalMessageId}
         values={{
           hrid: holdingsRecord.hrid,
-          location: holdingsPermanentLocation,
+          location: holdingsPermanentLocationName,
           itemCount,
         }}
       />
@@ -678,7 +692,13 @@ class ViewHoldingsRecord extends React.Component {
                   paneTitle={intl.formatMessage({
                     id: 'ui-inventory.holdingsPaneTitle',
                   }, {
-                    location: referenceTables?.locationsById[holdingsRecord?.effectiveLocationId]?.name,
+                    location:
+                      holdingsEffectiveLocation?.isActive ?
+                        holdingsEffectiveLocation?.name :
+                        intl.formatMessage(
+                          { id: 'ui-inventory.inactive.paneTitle' },
+                          { location: holdingsEffectiveLocation?.name }
+                        ),
                     callNumber: holdingsRecord?.callNumber,
                   })}
                   paneSub={intl.formatMessage({
@@ -820,13 +840,21 @@ class ViewHoldingsRecord extends React.Component {
                           >
                             <KeyValue
                               label={<FormattedMessage id="ui-inventory.permanent" />}
-                              value={checkIfElementIsEmpty(locationAccordion.permanent)}
+                              value={checkIfElementIsEmpty(locationAccordion.permanent?.name)}
+                              subValue={(!locationAccordion.permanent?.isActive) &&
+                                <FormattedMessage id="ui-inventory.inactive" />
+                              }
                             />
                           </Col>
                           <Col sm={4}>
                             <KeyValue
                               label={<FormattedMessage id="ui-inventory.temporary" />}
-                              value={checkIfElementIsEmpty(locationAccordion.temporary)}
+                              value={checkIfElementIsEmpty(locationAccordion.temporary?.name)}
+                              subValue={(locationAccordion.temporary &&
+                                !locationAccordion.temporary?.isActive) &&
+                                <FormattedMessage id="ui-inventory.inactive" />
+                              }
+                              data-test-id="temporary-location"
                             />
                           </Col>
                         </Row>
@@ -996,6 +1024,9 @@ class ViewHoldingsRecord extends React.Component {
                       />
 
                       <HoldingReceivingHistory holding={holdingsRecord} />
+
+                      <HoldingBoundWith boundWithItems={boundWithItems} />
+
                     </AccordionSet>
                   </AccordionStatus>
                 </Pane>
@@ -1024,6 +1055,9 @@ ViewHoldingsRecord.propTypes = {
       records: PropTypes.arrayOf(PropTypes.object),
     }),
     tagSettings: PropTypes.shape({
+      records: PropTypes.arrayOf(PropTypes.object),
+    }),
+    boundWithItems: PropTypes.shape({
       records: PropTypes.arrayOf(PropTypes.object),
     }),
   }).isRequired,